fix: resolve type erasure bug in Java behavior-mode instrumentation
Behavior-mode instrumentation captured function return values as Object,
causing compilation errors when substituted back into generic contexts
(e.g. List<Long>.add(func())). Fixed by extracting the target function's
return type via tree-sitter and casting the Object variable back to the
correct type when _infer_array_cast_type cannot determine it from context.

+++ tests/test_languages/test_java/test_instrumentation.py
@@ -2743,7 +2743,7 @@ public class CounterTest__perfinstrumented {
                 }
             }
         }
-        assertEquals(1, _cf_result1_1);
+        assertEquals(1, (int)_cf_result1_1);
     }
 }
 """
